refactor(sentry): extract init and add device_id tag
